refactor: move to nav group types
