+++ src/admin/components/elements/NavItem/types.ts
@@ -1,16 +1,5 @@
-import { SvgIconComponent } from '@mui/icons-material';
-
-export type Group = {
-  label: string;
-  items: GroupItem[];
-};
+import { GroupItem } from '../NavGroup/types.js';
 
 export type Props = {
   item: GroupItem;
-};
-
-export type GroupItem = {
-  label: string;
-  href: string;
-  Icon: SvgIconComponent;
 };
